Add edit button for Creatures
Made so the Creature passed to CreatureInitiativeItem is no longer a
prop. It is a v-model instead.

Made so that, when editing a Creature, CreatureForm disables some inputs
not intended to be modified. These include:
* `isPlayer`
* `number`: can modify only one at a time
* `hitPointsCurrent`: use the VNumberInput in CreatureInitiativeItem
  instead.

Fixed typings (null vs undefined).

As an unwanted side-effect, now CreatureForms starts with `isPlayer` to
false instead of true when opened from NewCreature. This is because it
is an optional v-model, and the default value of a boolean prop is false,
not undefined, so it cannot be recognized when it is not passed.